requirement.txt
numpy
opencv-python
keras-gpu
pillow
glob
itertools
